fixed delete

+++ library/views.py
@@ -128,11 +128,54 @@ def modify_pub(request, pub_id):
         publisher = get_object_or_404(Publisher, pk=pub_id)
         return render(request, 'forms/modify_publisher.html', {'publisher': publisher})
 
+
+def addbook(request):
+    if(request.method == 'POST'):
+        b_title = request.POST['b_title']
+        b_isbn = request.POST['b_isbn']
+        b_year = request.POST['b_year']
+        b_author = request.POST['b_author']
+        b_publisher = request.POST['b_publisher']
+
+        book = Book(title=b_title, isbn=b_isbn, year=b_year, author=b_author, publisher=b_publisher)
+        book.save()
+        
+        return redirect('/home/books')
+    else:
+        author = Author.objects.all()
+        publisher = Publisher.objects.all()
+        books = Book.objects.all()
+        context = {
+            'author': author,
+            'publisher': publisher,
+            'books': books,
+        }
+        
+        return render(request, 'forms/add_books.html', context)
+
+
+
+
+
 #delete author, and publisher views
 
 def delete_a(request, author_id):
-    a = Author.objects.get(pk=author_id)
-    a.delete()
-    return redirect('/home/author')
+    if(request.method == 'DELETE'):
+        try:
+            a = Author.objects.get(pk=author_id)
+        except Author.DoesNotExist:
+            return HttpResponse("Author does not exist")
+        a.delete()
+        return redirect('/home/author') 
 
+def delete_p(request, pub_id):
+    if(request.method == 'DELETE'):
+        try:
+            p = Publisher.objects.get(pk=pub_id)
+        except Publisher.DoesNotExist:
+            return HttpResponse("Publisher does not exist")
+        p.delete()
+        return redirect('/home/publisher') 
+        
+
